requirements.txt
# ===== Module 1: 爬蟲 =====
feedparser>=6.0.11
httpx>=0.27.0
aiohttp>=3.9.5
beautifulsoup4>=4.12.3
lxml>=5.2.2
playwright>=1.45.0
python-dateutil>=2.9.0
tldextract>=5.1.2

# ===== LLM (Claude) =====
anthropic>=0.34.0

# ===== 資料庫 =====
SQLAlchemy>=2.0.30

# ===== 設定 / 環境 =====
PyYAML>=6.0.1
python-dotenv>=1.0.1

# ===== 日誌 =====
loguru>=0.7.2

# === 以下模組在後續 Module 建置時才會用到，先列出 ===
# Module 2 (Brief + Telegram)
python-telegram-bot>=21.3
flask>=3.0.3

# Module 4 (TTS)
# TTS>=0.22.0            # Coqui XTTS-v2
# pydub>=0.25.1
# noisereduce>=3.0.2
# elevenlabs>=1.5.0

# Module 5 (ComfyUI client)
# websocket-client>=1.8.0

# Module 6 (影片 / 字幕)
# openai-whisper>=20240930
# Pillow>=10.4.0
# opencv-python>=4.10.0
# ffmpeg-python>=0.2.0

# Module 7 (YouTube 上傳)
# google-api-python-client>=2.140.0
# google-auth-oauthlib>=1.2.1

# Module 9 (排程)
# APScheduler>=3.10.4
